feat: API Access Token - Delete User

+++ src/pages/api/[[...user]].tsx
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
 import type { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(
   req: NextApiRequest,
@@ -41,21 +42,39 @@ export default async function handler(
     });
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    // jalankan delete data dari service firebase
-    await deleteData("users", user[1], (result: boolean) => {
-      if (result) {
-        res.status(200).json({
-          status: true,
-          statusCode: 200,
-          message: "Delete User Success",
-        });
-      } else {
-        res.status(400).json({
-          status: false,
-          statusCode: 400,
-          message: "Delete User Failed",
-        });
+    const token: any = req.headers.authorization?.split(" ")[1] ?? "";
+
+    // verifikasi token
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET ?? "",
+      async (err: any, decoded: any) => {
+        // jika decoded tidak null dan role sebagai admin, maka bisa delete user
+        if (decoded && decoded.role === "admin") {
+          // jalankan delete data dari service firebase
+          await deleteData("users", user[1], (result: boolean) => {
+            if (result) {
+              res.status(200).json({
+                status: true,
+                statusCode: 200,
+                message: "Delete User Success",
+              });
+            } else {
+              res.status(400).json({
+                status: false,
+                statusCode: 400,
+                message: "Delete User Failed",
+              });
+            }
+          });
+        } else {
+          res.status(403).json({
+            status: false,
+            statusCode: 403,
+            message: "Access Denied!",
+          });
+        }
       }
-    });
+    );
   }
 }
